fix xcode warnings

+++ NRDB/CardImagePopup.m
@@ -80,8 +80,8 @@ static UIPopoverController* popover;
     self.copiesLabel.textColor = [UIColor blackColor];
     if ([self.cc.card.setCode isEqualToString:@"core"])
     {
-        int cores = [[NSUserDefaults standardUserDefaults] integerForKey:NUM_CORES];
-        int owned = cores * self.cc.card.quantity;
+        NSInteger cores = [[NSUserDefaults standardUserDefaults] integerForKey:NUM_CORES];
+        NSInteger owned = cores * self.cc.card.quantity;
         
         if (owned < self.cc.count)
         {
